Moved madminer.utils.analysis functions to madminer.analysis class
